# Conflicts:
#	README.md
#	doc/若依环境使用手册.docx
#	pom.xml
#	zzyl-admin/pom.xml
#	zzyl-admin/src/main/java/com/zzyl/RuoYiApplication.java
#	zzyl-admin/src/main/java/com/zzyl/web/controller/common/CommonController.java
#	zzyl-admin/src/main/java/com/zzyl/web/controller/system/SysLoginController.java
#	zzyl-admin/src/main/java/com/zzyl/web/controller/system/SysProfileController.java
#	zzyl-admin/src/main/java/com/zzyl/web/controller/system/SysUserController.java
#	zzyl-admin/src/main/java/com/zzyl/web/core/config/SwaggerConfig.java
#	zzyl-admin/src/main/resources/application.yml
#	zzyl-common/pom.xml
#	zzyl-common/src/main/java/com/zzyl/common/constant/CacheConstants.java
#	zzyl-common/src/main/java/com/zzyl/common/constant/Constants.java
#	zzyl-common/src/main/java/com/zzyl/common/constant/GenConstants.java
#	zzyl-common/src/main/java/com/zzyl/common/core/domain/entity/SysUser.java
#	zzyl-common/src/main/java/com/zzyl/common/utils/file/FileUploadUtils.java
#	zzyl-common/src/main/java/com/zzyl/common/utils/file/FileUtils.java
#	zzyl-common/src/main/java/com/zzyl/common/utils/poi/ExcelUtil.java
#	zzyl-framework/pom.xml
#	zzyl-framework/src/main/java/com/zzyl/framework/config/FilterConfig.java
#	zzyl-framework/src/main/java/com/zzyl/framework/interceptor/MyMetaObjectHandler.java
#	zzyl-framework/src/main/java/com/zzyl/framework/web/service/SysLoginService.java
#	zzyl-framework/src/main/java/com/zzyl/framework/web/service/SysRegisterService.java
#	zzyl-generator/pom.xml
#	zzyl-generator/src/main/java/com/zzyl/generator/util/GenUtils.java
#	zzyl-generator/src/main/java/com/zzyl/generator/util/VelocityUtils.java
#	zzyl-generator/src/main/resources/generator.yml
#	zzyl-generator/src/main/resources/vm/java/controller.java.vm
#	zzyl-generator/src/main/resources/vm/java/domain.java.vm
#	zzyl-generator/src/main/resources/vm/java/mapper.java.vm
#	zzyl-generator/src/main/resources/vm/java/serviceImpl.java.vm
#	zzyl-nursing-platform/pom.xml
#	zzyl-nursing-platform/src/main/java/com/zzyl/nursing/controller/NursingLevelController.java
#	zzyl-nursing-platform/src/main/java/com/zzyl/nursing/controller/NursingPlanController.java
#	zzyl-nursing-platform/src/main/java/com/zzyl/nursing/controller/NursingProjectController.java
#	zzyl-nursing-platform/src/main/java/com/zzyl/nursing/domain/NursingLevel.java
#	zzyl-nursing-platform/src/main/java/com/zzyl/nursing/domain/NursingPlan.java
#	zzyl-nursing-platform/src/main/java/com/zzyl/nursing/domain/NursingProject.java
#	zzyl-nursing-platform/src/main/java/com/zzyl/nursing/domain/NursingProjectPlan.java
#	zzyl-nursing-platform/src/main/java/com/zzyl/nursing/dto/NursingPlanDto.java
#	zzyl-nursing-platform/src/main/java/com/zzyl/nursing/mapper/NursingLevelMapper.java
#	zzyl-nursing-platform/src/main/java/com/zzyl/nursing/mapper/NursingPlanMapper.java
#	zzyl-nursing-platform/src/main/java/com/zzyl/nursing/mapper/NursingProjectMapper.java
#	zzyl-nursing-platform/src/main/java/com/zzyl/nursing/mapper/NursingProjectPlanMapper.java
#	zzyl-nursing-platform/src/main/java/com/zzyl/nursing/service/INursingLevelService.java
#	zzyl-nursing-platform/src/main/java/com/zzyl/nursing/service/INursingPlanService.java
#	zzyl-nursing-platform/src/main/java/com/zzyl/nursing/service/INursingProjectService.java
#	zzyl-nursing-platform/src/main/java/com/zzyl/nursing/service/impl/NursingLevelServiceImpl.java
#	zzyl-nursing-platform/src/main/java/com/zzyl/nursing/service/impl/NursingPlanServiceImpl.java
#	zzyl-nursing-platform/src/main/java/com/zzyl/nursing/service/impl/NursingProjectServiceImpl.java
#	zzyl-nursing-platform/src/main/java/com/zzyl/nursing/vo/NursingPlanVo.java
#	zzyl-nursing-platform/src/main/resources/mapper/nursing/NursingLevelMapper.xml
#	zzyl-nursing-platform/src/main/resources/mapper/nursing/NursingPlanMapper.xml
#	zzyl-nursing-platform/src/main/resources/mapper/nursing/NursingProjectMapper.xml
#	zzyl-nursing-platform/src/main/resources/mapper/nursing/NursingProjectPlanMapper.xml
#	zzyl-oss/pom.xml
#	zzyl-oss/src/main/java/com/zzyl/oss/AliyunOSSOperator.java
#	zzyl-oss/src/main/java/com/zzyl/oss/AliyunOSSProperties.java
#	zzyl-quartz/pom.xml
#	zzyl-quartz/src/main/java/com/zzyl/quartz/util/ScheduleUtils.java
#	zzyl-system/pom.xml
#	zzyl-system/src/main/java/com/zzyl/system/mapper/SysUserMapper.java
#	zzyl-system/src/main/java/com/zzyl/system/service/ISysUserService.java
#	zzyl-system/src/main/java/com/zzyl/system/service/impl/SysUserServiceImpl.java
#	zzyl-system/src/main/resources/mapper/system/SysUserMapper.xml
#	zzyl-ui/package.json
#	zzyl-ui/src/components/RightToolbar/index.vue
#	zzyl-ui/src/layout/components/AppMain.vue
#	zzyl-ui/src/layout/components/Navbar.vue
#	zzyl-ui/src/layout/components/Settings/index.vue
#	zzyl-ui/src/layout/components/TagsView/index.vue
#	zzyl-ui/src/layout/index.vue
#	zzyl-ui/src/settings.js
#	zzyl-ui/src/store/modules/settings.js
#	zzyl-ui/src/store/modules/user.js
#	zzyl-ui/src/views/index.vue
#	zzyl-ui/src/views/system/user/index.vue
#	zzyl-ui/src/views/system/user/profile/index.vue

+++ zzyl-nursing-platform/src/main/java/com/zzyl/nursing/dto/NursingPlanDto.java
@@ -16,7 +16,7 @@ public class NursingPlanDto {
      * 排序号
      */
     @ApiModelProperty(value = "排序号")
-    private Long sortNo;
+    private Integer sortNo;
 
     /**
      * 计划名称
@@ -28,7 +28,7 @@ public class NursingPlanDto {
      * 状态（0：禁用，1：启用）
      */
     @ApiModelProperty(value = "状态（0：禁用，1：启用）")
-    private Long status;
+    private Integer status;
 
     /**
      * 护理计划关联项目列表
